+++ src/PairingPlus.cpp
@@ -1,7 +1,7 @@
 /**
  * @file src/PairingPlus.cpp
  *
- * @date 2015-05-06
+ * @date 2016-01-21
  *
  * @author Youri Hoogstrate
  *
@@ -38,12 +38,12 @@
 /**
  * @brief Stores positions and sets PairingType
  *
- * @date 2015-05-06
+ * @date 2016-01-21
  */
 PairingPlus::PairingPlus(Position arg_position1, Position arg_position2):
 	position1(arg_position1),
 	position2(arg_position2),
-	size(arg_position2 - arg_position1 - 1)
+	size((size_t)(arg_position2 - arg_position1 - 1))
 {
 	this->init();
 }
